refactor!: use `uint8` instead of `uint256` in LSP7 decimals (#292)

+++ contracts/LSP7DigitalAsset/LSP7DigitalAssetCore.sol
@@ -45,7 +45,7 @@ abstract contract LSP7DigitalAssetCore is ILSP7DigitalAsset {
     /**
      * @inheritdoc ILSP7DigitalAsset
      */
-    function decimals() public view override returns (uint256) {
+    function decimals() public view override returns (uint8) {
         return _isNonDivisible ? 0 : 18;
     }
